# 8307782: Compensate for timed wait lags on Thread.sleep paths