The underlying Thread.sleep facilities consistently overshoot the requested sleep durations. Part of that are issues like [JDK-8307766](https://bugs.openjdk.org/browse/JDK-8307766), but there are also generic overhead from the various scheduling/monitoring/accounting paths in the kernel itself.

We can improve this a bit by accounting for the lags straight in Thread.sleep code paths.

Motivational improvements on `c6.8xlarge`, the sleep accuracy is improved the following way:

```
% CONF=linux-x86_64-server-release make test TEST=micro:ThreadSleep.millisNanos

Benchmark                   (sleep)  Mode  Cnt           Score      Error  Units

# Baseline
ThreadSleep.millisNanos           0  avgt   15         473.727 ±    3.542  ns/op
ThreadSleep.millisNanos           1  avgt   15         585.868 ±    2.010  ns/op
ThreadSleep.millisNanos          10  avgt   15         583.475 ±    1.426  ns/op
ThreadSleep.millisNanos         100  avgt   15       57833.073 ±  146.406  ns/op
ThreadSleep.millisNanos        1000  avgt   15       59324.901 ±  620.924  ns/op
ThreadSleep.millisNanos       10000  avgt   15       68071.990 ±  537.880  ns/op
ThreadSleep.millisNanos      100000  avgt   15      158069.170 ±  329.991  ns/op
ThreadSleep.millisNanos     1000000  avgt   15     1059044.770 ±  370.982  ns/op
ThreadSleep.millisNanos    10000000  avgt   15    10060017.354 ±  399.309  ns/op
ThreadSleep.millisNanos   100000000  avgt   15   100062969.073 ± 2399.053  ns/op
ThreadSleep.millisNanos  1000000000  avgt   15  1000068474.467 ± 2669.980  ns/op

# Patched
ThreadSleep.millisNanos           0  avgt   15         473.547 ±     3.091  ns/op
ThreadSleep.millisNanos           1  avgt   15         584.989 ±     1.643  ns/op
ThreadSleep.millisNanos          10  avgt   15         582.831 ±     0.526  ns/op
ThreadSleep.millisNanos         100  avgt   15         584.225 ±     1.922  ns/op
ThreadSleep.millisNanos        1000  avgt   15       58212.649 ±   399.776  ns/op
ThreadSleep.millisNanos       10000  avgt   15       60571.290 ±   441.015  ns/op
ThreadSleep.millisNanos      100000  avgt   15      158212.772 ±   480.856  ns/op
ThreadSleep.millisNanos     1000000  avgt   15     1016383.173 ±  7825.281  ns/op
ThreadSleep.millisNanos    10000000  avgt   15    10009288.895 ±  3963.795  ns/op
ThreadSleep.millisNanos   100000000  avgt   15   100020406.367 ± 10422.899  ns/op
ThreadSleep.millisNanos  1000000000  avgt   15  1000030401.933 ± 18589.040  ns/op
```

Together with [JDK-8307766](https://bugs.openjdk.org/browse/JDK-8307766) it improves the precision even more:

```
Benchmark                   (sleep)  Mode  Cnt           Score      Error  Units

# -XX:TimerSlack=1, Baseline
ThreadSleep.millisNanos           0  avgt   15         469.632 ±    1.500  ns/op
ThreadSleep.millisNanos           1  avgt   15         587.562 ±    8.966  ns/op
ThreadSleep.millisNanos          10  avgt   15         584.241 ±    1.997  ns/op
ThreadSleep.millisNanos         100  avgt   15        7907.160 ±   12.476  ns/op
ThreadSleep.millisNanos        1000  avgt   15        7913.984 ±   56.996  ns/op
ThreadSleep.millisNanos       10000  avgt   15       17929.107 ±  432.083  ns/op
ThreadSleep.millisNanos      100000  avgt   15      108154.851 ±  403.717  ns/op
ThreadSleep.millisNanos     1000000  avgt   15     1008854.095 ±  418.444  ns/op
ThreadSleep.millisNanos    10000000  avgt   15    10009902.829 ±  279.617  ns/op
ThreadSleep.millisNanos   100000000  avgt   15   100014000.020 ± 2143.820  ns/op
ThreadSleep.millisNanos  1000000000  avgt   15  1000020734.333 ± 4846.754  ns/op

# -XX:TimerSlack=1, Patched
ThreadSleep.millisNanos           0  avgt   15         472.145 ±    2.440  ns/op
ThreadSleep.millisNanos           1  avgt   15         584.948 ±    0.182  ns/op
ThreadSleep.millisNanos          10  avgt   15         583.426 ±    1.265  ns/op
ThreadSleep.millisNanos         100  avgt   15         585.094 ±    0.775  ns/op
ThreadSleep.millisNanos        1000  avgt   15        7886.877 ±   56.901  ns/op
ThreadSleep.millisNanos       10000  avgt   15       13184.087 ± 1080.463  ns/op
ThreadSleep.millisNanos      100000  avgt   15      100838.978 ±   99.685  ns/op
ThreadSleep.millisNanos     1000000  avgt   15     1002433.084 ±  264.073  ns/op
ThreadSleep.millisNanos    10000000  avgt   15    10002420.847 ±  621.984  ns/op
ThreadSleep.millisNanos   100000000  avgt   15   100006729.862 ± 3303.521  ns/op
ThreadSleep.millisNanos  1000000000  avgt   15  1000011740.600 ± 3688.820  ns/op
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [ ] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8307782](https://bugs.openjdk.org/browse/JDK-8307782): Compensate for timed wait lags on Thread.sleep paths


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/13892/head:pull/13892` \
`$ git checkout pull/13892`

Update a local copy of the PR: \
`$ git checkout pull/13892` \
`$ git pull https://git.openjdk.org/jdk.git pull/13892/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 13892`

View PR using the GUI difftool: \
`$ git pr show -t 13892`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/13892.diff">https://git.openjdk.org/jdk/pull/13892.diff</a>

</details>